# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/offices": {
            "get": {
                "description": "list office entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List office entities",
                "operationId": "list-office",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Office"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create office",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create office",
                "operationId": "create-office",
                "parameters": [
                    {
                        "description": "Office entity",
                        "name": "office",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Office"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Office"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/offices/{id}": {
            "get": {
                "description": "get office by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a office entity by ID",
                "operationId": "get-office",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Office ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Office"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get office by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a office entity by ID",
                "operationId": "delete-office",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Office ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ent.Department": {
            "type": "object",
            "properties": {
                "Name": {
                    "description": "Name holds the value of the \"Name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DepartmentQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DepartmentEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.DepartmentEdges": {
            "type": "object",
            "properties": {
                "details": {
                    "description": "Details holds the value of the details edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Detail"
                    }
                },
                "offices": {
                    "description": "Offices holds the value of the offices edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Office"
                    }
                },
                "schedules": {
                    "description": "Schedules holds the value of the schedules edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Schedule"
                    }
                },
                "specialdoctors": {
                    "description": "Specialdoctors holds the value of the specialdoctors edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Specialdoctor"
                    }
                },
                "trainings": {
                    "description": "Trainings holds the value of the trainings edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Training"
                    }
                }
            }
        },
        "ent.Detail": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DetailQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DetailEdges"
                },
                "explain": {
                    "description": "Explain holds the value of the \"explain\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.DetailEdges": {
            "type": "object",
            "properties": {
                "course": {
                    "description": "Course holds the value of the course edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Course"
                },
                "department": {
                    "description": "Department holds the value of the department edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Department"
                },
                "mission": {
                    "description": "Mission holds the value of the mission edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Mission"
                }
            }
        },
        "ent.Disease": {
            "type": "object",
            "properties": {
                "disease": {
                    "description": "Disease holds the value of the \"disease\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DiseaseQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DiseaseEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.DiseaseEdges": {
            "type": "object",
            "properties": {
                "doctors": {
                    "description": "Doctors holds the value of the doctors edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Doctor"
                    }
                }
            }
        },
        "ent.Doctor": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "Address holds the value of the \"address\" field.",
                    "type": "string"
                },
                "age": {
                    "description": "Age holds the value of the \"age\" field.",
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DoctorQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DoctorEdges"
                },
                "educational": {
                    "description": "Educational holds the value of the \"educational\" field.",
                    "type": "string"
                },
                "email": {
                    "description": "Email holds the value of the \"email\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "password": {
                    "description": "Password holds the value of the \"password\" field.",
                    "type": "string"
                },
                "phone": {
                    "description": "Phone holds the value of the \"phone\" field.",
                    "type": "string"
                }
            }
        },
        "ent.DoctorEdges": {
            "type": "object",
            "properties": {
                "departments": {
                    "description": "Departments holds the value of the departments edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Department"
                    }
                },
                "disease": {
                    "description": "Disease holds the value of the disease edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Disease"
                },
                "gender": {
                    "description": "Gender holds the value of the gender edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Gender"
                },
                "offices": {
                    "description": "Offices holds the value of the offices edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Office"
                    }
                },
                "position": {
                    "description": "Position holds the value of the position edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Position"
                },
                "schedules": {
                    "description": "Schedules holds the value of the schedules edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Schedule"
                    }
                },
                "specialdoctors": {
                    "description": "Specialdoctors holds the value of the specialdoctors edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Specialdoctor"
                    }
                },
                "title": {
                    "description": "Title holds the value of the title edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Title"
                },
                "trainings": {
                    "description": "Trainings holds the value of the trainings edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Training"
                    }
                }
            }
        },
        "ent.Extradoctor": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ExtradoctorQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ExtradoctorEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "specialname": {
                    "description": "Specialname holds the value of the \"specialname\" field.",
                    "type": "string"
                }
            }
        },
        "ent.ExtradoctorEdges": {
            "type": "object",
            "properties": {
                "offices": {
                    "description": "Offices holds the value of the offices edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Office"
                    }
                },
                "specialdoctors": {
                    "description": "Specialdoctors holds the value of the specialdoctors edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Specialdoctor"
                    }
                }
            }
        },
        "ent.Gender": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the GenderQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.GenderEdges"
                },
                "gender": {
                    "description": "Gender holds the value of the \"gender\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.GenderEdges": {
            "type": "object",
            "properties": {
                "doctors": {
                    "description": "Doctors holds the value of the doctors edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Doctor"
                    }
                }
            }
        },
...
        "ent.Office": {
            "type": "object",
            "properties": {
                "added_time1": {
                    "description": "AddedTime1 holds the value of the \"added_time1\" field.",
                    "type": "string"
                },
                "added_time2": {
                    "description": "AddedTime2 holds the value of the \"added_time2\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the OfficeQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.OfficeEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "officename": {
                    "description": "Officename holds the value of the \"officename\" field.",
                    "type": "string"
                },
                "specialdoctor_id": {
                    "type": "integer"
                }
            }
        },
        "ent.OfficeEdges": {
            "type": "object",
            "properties": {
                "department": {
                    "description": "Department holds the value of the department edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Department"
                },
                "doctor": {
                    "description": "Doctor holds the value of the doctor edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Doctor"
                },
                "extradoctor": {
                    "description": "Extradoctor holds the value of the extradoctor edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Extradoctor"
                },
                "schedules": {
                    "description": "Schedules holds the value of the schedules edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Schedule"
                    }
                }
            }
        },
        "ent.Position": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PositionQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PositionEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "position": {
                    "description": "Position holds the value of the \"position\" field.",
                    "type": "string"
                }
            }
        },
        "ent.PositionEdges": {
            "type": "object",
            "properties": {
                "doctors": {
                    "description": "Doctors holds the value of the doctors edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Doctor"
                    }
                }
            }
        },
        "ent.Schedule": {
            "type": "object",
            "properties": {
                "activity": {
                    "description": "Activity holds the value of the \"activity\" field.",
                    "type": "string"
                },
                "added_time": {
                    "description": "AddedTime holds the value of the \"added_time\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ScheduleQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ScheduleEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.ScheduleEdges": {
            "type": "object",
            "properties": {
                "department": {
                    "description": "Department holds the value of the department edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Department"
                },
                "docter": {
                    "description": "Docter holds the value of the docter edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Doctor"
                },
                "office": {
                    "description": "Office holds the value of the office edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Office"
                }
            }
        },
        "ent.Specialdoctor": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the SpecialdoctorQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.SpecialdoctorEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "other": {
                    "description": "Other holds the value of the \"other\" field.",
                    "type": "string"
                }
            }
        },
        "ent.SpecialdoctorEdges": {
            "type": "object",
            "properties": {
                "department": {
                    "description": "Department holds the value of the department edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Department"
                },
                "doctor": {
                    "description": "Doctor holds the value of the doctor edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Doctor"
                },
                "extradoctor": {
                    "description": "Extradoctor holds the value of the extradoctor edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Extradoctor"
                },
                "offices": {
                    "description": "Offices holds the value of the offices edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Office"
                    }
                }
            }
        },
        "ent.Title": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the TitleQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.TitleEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "title": {
                    "description": "Title holds the value of the \"title\" field.",
                    "type": "string"
                }
            }
        },
        "ent.TitleEdges": {
            "type": "object",
            "properties": {
                "doctors": {
                    "description": "Doctors holds the value of the doctors edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Doctor"
                    }
                }
            }
        },
        "ent.Training": {
            "type": "object",
            "properties": {
                "branch": {
                    "description": "Branch holds the value of the \"branch\" field.",
                    "type": "string"
                },
                "dateone": {
                    "description": "Dateone holds the value of the \"dateone\" field.",
                    "type": "string"
                },
                "datetwo": {
                    "description": "Datetwo holds the value of the \"datetwo\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the TrainingQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.TrainingEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.TrainingEdges": {
            "type": "object",
            "properties": {
                "course": {
                    "description": "Course holds the value of the course edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Course"
                },
                "department": {
                    "description": "Department holds the value of the department edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Department"
                },
                "doctor": {
                    "description": "Doctor holds the value of the doctor edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Doctor"
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "authorizationUrl": "",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "authorizationUrl": "",
...
}